new security_checker facet
Summary:
Until now we had 3 different ways of permission checking in Mononoke across our
3 surfaces (scs server, writeproto and eden api). Let's start unifying them by
moving the wireproto logic to from repo_listener to its own facet.

In this diff I'm moving the mononoke-server logic to the new facet as-is (that
is without changing the logic. In the next diffs I will migrate other monoonke
ACL checks to use it.

Reviewed By: HarveyHunt

Differential Revision: D35178977

fbshipit-source-id: befbcb5e78482efc09558838846cc20c76655ae7