Implement logic for into_entries
Summary:
**Context:** Sharded maps will be used on deleted manifest v2

This diff adds the logic for the `into_entries` operation on sharded maps. That's the equivalent to `.into_iter` on a regular map.

We take special care to only load things in memory as needed, which should mean we never need to load more than a couple of blobs in memory at the same time.

Reviewed By: kris1319

Differential Revision: D35251932

fbshipit-source-id: 0005b34f254180afd78d112a34b2de067f968e80