Implement logic for update
Summary:
**Context:** Sharded maps will be used on deleted manifest v2

This diff adds the implementation for the `update` function, which generalises all write operations (adding and removing elements). This is the meat of all the "improved deleted manifest" stuff, most of the other things on top will just be adaptors and boilerplate code.

As it generalises both operations, it needs to deal with all the details of keeping the optimisations in place, plus we also need to be careful so that the structure is history-independent (which should be the case if we keep use the optimisations correctly).

This is a complicated diff, and I recommend also reading [this explanation](https://fburl.com/lnusbzgl) to understand the code. I'm also happy to set a 1:1 and whiteboard the logic to make it clearer.

Reviewed By: markbt

Differential Revision: D35288510

fbshipit-source-id: fc2c1eb0f92dff7c131c4e742007886b0d01196a