Do for concurrently for efficiency
Summary:
On this part of the algorithm, we do a bunch of serial awaits, but they don't need to be serial, we can run them together.

 ---

Partly related, I am also considering changing `lookup` function to take a vec of things to lookup so we can look them up together more efficiently, which would help here (specially on DMv2), but it might be a bit overkill as cache should already help a lot here.

Reviewed By: markbt

Differential Revision: D35399538

fbshipit-source-id: 19902c2881edb2e078a390972b7e1f7fe3298d86