Store size on edge, not node
Summary:
This changes sharded maps so the size is stored on the edge, not on the node itself. See [1] for more discussion.

This has a few advantages:
- No need to load the node to know the size.
- This makes it possible to do some optimisations, see [2]
- Since we can easily look at the sizes of children, we don't need to be super extra careful to keep the correct size while updating the map, and can just calculate it at the end. See [3] where code is a little cleaner.

And disadvantages:
- Code is a little more boilerplaty, as it needs yet another struct to hold the additional size.
- To calculate the size of the current node, we need to traverse the children. That should not be a big deal as we have to load them anyway. I also put it behind a OnceCell so it's not inefficient.

IMO I'm fine with going either way. I think size on edge is a little uglier to code but a little more practical, so I prefer it. But I'm fine to go with size-on-node if someone else has a stronger opinion.

markbt has thought about this and might have opinions. I don't think using a separate node is good as this is only for size (which is necessary for the algorithm to be fast, not "aggregated data").

Reviewed By: markbt

Differential Revision: D35314803

fbshipit-source-id: 13f488dbbc5cf07a82fac9d0e638682c7d9fd926