init: remove the use of init.use-rust config for inner Rust repo object loading
Summary:
Recently, a check was added for avoiding loading the inner Rust repo if `init.use-rust` was not enabled. This was a mitigation for users seeing excessive reloads in the reponame file caused by the reload of the dynamic config, which was caused by the loading of the inner Rust repo object.

This also should be fine since we are rolling that config to users this week anyways.

Reviewed By: quark-zju

Differential Revision: D35300521

fbshipit-source-id: bba2e37217f8bd3cdd11c6091c404389ec3c3b2b