repo: add metalog into Rust repo
Summary: Adds a couple of methods for making the Rust repo object handle the metalog, and makes use of this on the Python side. Since the way the metalog is used in Python is tightly coupled with the VFS object, trying to make use direct of it would be rather complicated, so we are still going to use it through the VFS object.

Reviewed By: quark-zju

Differential Revision: D35256084

fbshipit-source-id: 273ad118ebbeb72e04f43020e2c48cab4958d162