repo: add EdenAPI into Rust repo
Summary:
Adds a method for making the Rust repo object hold the EdenAPI, and uses it for replacing `edenapi.getclient`.

Since most of the config checks done on the Python side don't seem to be useful in the new Rust world (most of them were used for testing), porting them or adding them to the bindings was simply not done.

Differential Revision: D35295652

fbshipit-source-id: 3058b02f2bb18d5a9b0697b0114e2c966fef5b73