testtool: move write_stub_log_entry into testtool
Summary:
Migrate the specialist `write_stub_log_entry` test tool into a subcommand of `mononoke_testtool`.

The existing tool is a bit of a misnomer: while it's purpose is to add entries into the bookmark update log, it actually does so no by writing stub entries, but by actually moving the bookmark.   So let's name the subcommand what it is: `modify-bookmark`.

Reviewed By: kris1319

Differential Revision: D35395744

fbshipit-source-id: 3aff8acfd65090559ca90a4e3f89f2806586a5f9