purge: don't do --dirs by default
Summary:
Previously "hg purge" would do "hg purge --files --dirs". However, that is slow since there is no fast way to find all empty directories to satisfy the "--dirs". Also, this behavior was somewhat surprising because the docs suggest --dirs is not included by default. Change the behavior to default to just --files (but add a config knob so we can revert this behavior quickly if needed).

Empty directories are normally invisible, and my presumption is very rarely the user cares about empty dirs when purging.

Reviewed By: yancouto

Differential Revision: D35515752

fbshipit-source-id: 3d4ab97309de1a9ef82987486fbbd1dbea3b91d8