mononoke][dm] Add method to derive deleted manifest for a simple stack
Summary:
This is basically a port of D30989888 (https://github.com/facebookexperimental/eden/commit/35a7998c07ceb61cb977a147017eb2c7bdaff9c6) to deleted manifests, and enables batch derivation for them. It does not have the limitation that commits can't delete files and turn them into directories.

The basic idea here is to parallelise the derivation in the tree, not in the commits. I added a comment explaining this idea in code, see the diff's contents.

This diff adds just a helper function for that, which I will integrate into the usual batch derivation trait in a following diff.

I didn't port the optimisations in the non-batch code: Batching some writes to blobstore and using a local mutexed hashmap to avoid double writes. I will add them in a following diff if unless I disprove their efficiency.

Reviewed By: farnz

Differential Revision: D35426790

fbshipit-source-id: 7432617737323a3e3854f4ba347bc86e8043ab2e