config: add reset state in SNAPSHOT
Summary:
In order to properly support reset on Windows, EdenFS needs to keep track of
both the checked out revision, and the reset one. Since this state needs to
persist across EdenFS restarts, we'll need to store it in the SNAPSHOT file.

This diff merely adds support for this new SNAPSHOT format, but doesn't write
it.

Reviewed By: chadaustin

Differential Revision: D35448641

fbshipit-source-id: 59fbafdadc885dfec8efd119c4fe1ff582987af0