sparse: add a test to prevent debugsparsematch regressions
Summary:
I'd like to use debugsparsematch in
[validate_sparse_profiles.py](https://www.internalfb.com/code/fbsource/tools/scm/sparse/arvr/.castle/validate_sparse_profiles.py)
script. The reason for doing that is because it has it's own sparse matcher
logic, which does not support sparse v2, which in turn causes failures on diffs
like D35461391.

Instead of reimplementing sparse match logic I'd suggest to just shell out to
hg (FWIW, we are already shelling out for buck in this script).

My biggest concern is that `debugsparsematch` is a debug command, and so in
theory it's format and/or input arguments can be changed and break
validate_sparse_profiles script. To prevent that I've added a small comment

Differential Revision: D35552480

fbshipit-source-id: cb781d749a7bff2eb8e6d4fcd234962eccc16970