mononoke: create usable segmented changelog in TestRepoFactory
Summary:
Currently unit tests that need a segmented changelog create it separately from the repo because in test repo there is only a `DisabledSegmentedChangelog`. I am making it possible to have an `OnDemandUpdateSegmentedChangelog` that is both useful and can be updated to include needed commits.

Look on new `test_is_ancestor` test to see what is made possible.

Reviewed By: markbt

Differential Revision: D35053513

fbshipit-source-id: 221600b50c8a0140b08f633dc3758489f2524f60