snapshot: Fix max-untracked-size with symlinks and set default to 1GB
Summary:
We currently have no default. Untracked files can be very large (like binaries and such), and easy to miss. We already had an option for that, but were not using it by default. Should the default be even smaller?

It also didn't work with invalid symlinks, as `Path.stat` tried to follow symlinks, so I fixed that.

If a client really wants to upload a large file, they can specify the command line argument by themselves.

Reviewed By: farnz

Differential Revision: D35498620

fbshipit-source-id: 7c21c779a680fb209b82943ec59aa5523cd1608c