avoid reloading segmented changelog if tailer did not make any progress
Summary:
We should not reload and reset the server-side segmented changelog if the
tailer has not written a new version since we last loaded - instead, we should
keep our existing one in-memory, because it will be more up to date than the
one we load.

In situations when the tailer is not working properly this might mitigate the issue for the customers instead of exposing them to it.

Reviewed By: farnz

Differential Revision: D35408139

fbshipit-source-id: 10d30b6d2d055fc938f9d0f22f3c633f7d7495c2