testing: initial transformer from .t to Python
Summary:
To be able to execute .t test without bash, and to be future
proof so tests can use Python more easily instead of bash.
Define a way that supports both classic .t and pure Python
and a mix of them.

See the docstring for details.

Not all `.t` features are handled in this diff.
Namely `#testcases` support is missing.

Reviewed By: LynBusch

Differential Revision: D34725133

fbshipit-source-id: f68c88a2e5c0fa4f3ce3de0b39ff17e11350b1be