testing: add extensions to extend TestTmp
Summary:
TestTmp itself is independent from application logic (ex. hg) and is
intended to be maybe useful for non-hg use-cases.

In hg's use-cases the hg and python command, and a bunch of env vars
like HGUSER need to be set.

Add an "extension" concept: `testsetup(t: TestTmp)` to extend TestTmp.
For now, just add hg and python (the command) as optional extensions.
In the future other users of the testing library might want to define
their own extensions.

Reviewed By: DurhamG

Differential Revision: D34790207

fbshipit-source-id: 8d5aac9a1fb5ad56d9df1c16a5240f77db6499b1