Rename/move InodeLoader tests to InodeOrTreeOrEntryLoader
Summary: `InodeLoader` is unused now. Delete it, and port it's tests over to `InodeOrTreeOrEntryLoader`. The test used to compare paths to verify results, because it was easy to walk the inode tree in reverse order to build up a path. We no longer have the path available, as Trees don't below to a hard parent (they can be moved, similar to inodes). Instead, we now set the contents of the file to the absolute path, and compare `Hash20` objects to verify we got the file we expected.

Reviewed By: xavierd

Differential Revision: D35663897

fbshipit-source-id: a58458a7d69aae283dfad031214344f02ebba850