dump-changesets: Change subcommand to dump-changesets and allow not fetching public changesets
Summary:
This command is also useful to just change format of dumped changesets (for example, I dumped them thrift serialized and used that for backfilling, but now need it plaintext for validation). In this case there's no "fetching public" at all, so I changed the command to be a bit more generic than that, and fetching public changesets being only an option on it.

The new way to structure args makes it a bit awkward to work with, as it's a bit hard to tell which args go in which part of the command, since some are common and some are not.

Reviewed By: mitrandir77

Differential Revision: D35580371

fbshipit-source-id: 4f32ed5af048642d72413d826eb3ee8b39cb494e