testlib: set production configs during setup
Summary:
One of the goals of the new test framework is to run tests with mostly
production configuration. Let's add the appropriate configs to the global hgrc
during test setup.

In the medium term we should get rid of this hard coded list and use
dynamicconfig to generate configs for the tests. But for now let's go with the
hard coded list.

The configs were mostly copied from library.sh a tinit.sh.

Differential Revision: D35683766

fbshipit-source-id: d802066bc801165f361c3418d51ae0e9181629c2