Prevent history cycles in mutable renames
Summary:
Making blame work in the presence of history cycles is non-trivial to get right, and my efforts have led to repeated infinite loops in SCS.

Simply banning cyclic history removes the problem - for now, use an approximation to get it right quickly, and make the problem go away.

Note that the test tests for cases that could be made to work - I hope that this is sufficient that code review will catch badly thought through efforts to make the check precise.

Reviewed By: yancouto

Differential Revision: D35810936

fbshipit-source-id: 19f5f3221d36e6c4270748495dd7de6d45e61ff8