mononoke: lfs: Return filestore upload result from internal_upload
Summary:
When uploading content to the LFS server it attempts to store the
content in an upstream server, as well as internal data stores. If either of
those fail, then the upload should be considered a failure and the client
should retry. However, D22192005 (https://github.com/facebookexperimental/eden/commit/ce7f53422f79736ab2ca6ac9293c881e1422d6d8) refactored the code and accidentally updated
`internal_upload` to always return `Ok(())`.

This meant that a failure to write to the filestore wasn't propagated and so an
upload that actually failed would send an HTTP 200 to the client.
Unfortunately, when the client asks for the blob later the server doesn't know
about it.

Update the code to return `res`, so that errors are propagated.

Reviewed By: krallin

Differential Revision: D35892336

fbshipit-source-id: 5302a34fdd338975e0f0db369ed66692db0ade9f