make segmented changelog manager a facet
Summary:
The repo factory is the place where we have all elements needed for
instantiating a segmented changelog manager. Let's add that as an option.

It's going to be useful for some specialized binaries that don't just want to
query the segmented changelog but want to understand what version of SC do we
currently store an to be able to reload SC at will.

Reviewed By: quark-zju

Differential Revision: D35744722

fbshipit-source-id: 37f9dc4b5750b48212be1bffbdae189341fb2a06