link subcommand for github extension
Summary:
With the `github` extension enabled, `hg link` associates pull request data
with a commit in the metalog. Note that nothing verifies that this
linkage exists on GitHub. Ultimately, `hg submit` will be required
to remap the commits within the actual pull request on GitHub.

I'm not sure what best practices are in terms of using flags
versus positional arguments in the CLI, but we can continue
to experiment with that, for now.

Reviewed By: quark-zju

Differential Revision: D35638199

fbshipit-source-id: f72514b13627a8ef845ffb99e6ae3c86098061cd