change E-Menu tooltip during checkout
Summary:
change the tooltip during a checkout operation. Also change it back when the checkout finishes.

The debug menu can also enable this tooltip by changing the icon color to orange. Changing the color to green or white will change the tooltip back

A similar tooltip is also implemented for when EdenFS detects an unhealthy mount

Reviewed By: kmancini

Differential Revision: D35488235

fbshipit-source-id: a33ebfd15cbfd0d46eea85f32be4ddd1bb09f357