pass job type to seed heads configuration
Summary:
We want to be able to return different heads for segmented changelog for
backgorund jobs vs live server. The seedheads function needs to be aware of
that distinction.

For now all SC abstraction generated by repo factor will be for live jobs
because they're used there (the tailer and repo imported construct the SC
explicitly).

Reviewed By: farnz

Differential Revision: D35934779

fbshipit-source-id: 732027d1db5e5f31237879dded3eed6d651edbe9