Implementation of commit_sparse_profile_size
Summary:
Commands calculates total size of profiles (all or provided list) by walking through the repo, matching paths to sparse profile config and using fsnodes info to get the size of files/directories.
Since in the directory with the sparse profile configs we have other files as well (validators configs, readme), I've decided to skip profile config if it is in bad format and just log it.

It took almost 27 minutes to process the request. In the next diffs I'm going to implement method to calculate delta size of profiles between 2 commits and store the all profile sizes for public commits.

Reviewed By: mitrandir77

Differential Revision: D35293351

fbshipit-source-id: 011f2e6144e8c35da226285296c9faae48a53e39