windows: teach eden to look for pdb in the binary path
Summary:
On Windows, dbghelp.dll does not look at the executable path for debug symbol files by default. This has caused us to not be able to see stack traces when iterating / debugging EdenFS on Windows.

This diff fixes it by manually adding the binary directory to the search list, and explicitly load it when reporting a crash.

This will also trigger loading from the debug symbol server **if it is configured in the environment variable**.

Reviewed By: xavierd

Differential Revision: D35623865

fbshipit-source-id: 72822bf38af93632fbd3a20757863d6bea4b078a