treeoverlay: set upper limit for journal size
Summary:
We often get reports complaining `treeoverlay.db-WAL` is too big on user's machine.

The issue here is that, we never get a chance to let SQLite run wal checkpoint because we never close the connection to SQLite database.

In which case, we can run the maintenance job ourselves to keep the journal file small. This diff adds a periodical task in EdenServer to give tree overlay a chance to run `wal_checkpoint`.

Reviewed By: xavierd

Differential Revision: D35321408

fbshipit-source-id: ac900be288d0919785876b93a9bde2166871c56a