inodes: fix overly aggressive XDCHECK
Summary:
On a case insensitive mount, changing the case of a directory is done by
removing the old directory first, then by adding the new directory. Thus
asserting that the new entry must not be a file is too aggressive, let's change
it to only fire on case insensitive mounts.

Reviewed By: genevievehelsel

Differential Revision: D36085219

fbshipit-source-id: ca5767cef48245b8bc1fdb314fe4c10222638a30