Add comments explaining use of BoxFuture
Summary:
On the traversal algorithms, we use BoxFuture on the types in order to make things more efficient.

That is very easy to miss and try to "fix it" like I did in D35848173, so this diff is adding comments explaining this. I think it is necessary to explain it because it is not clear from context, I thought we were using BoxFuture just because generics back in the past weren't strong enough.

Reviewed By: farnz

Differential Revision: D36131287

fbshipit-source-id: 7f741098387433bd5c73f2d30b7bd3f4785701b7