Add deleted_manifest_version and allow using DMv2 on Mononoke
Summary:
This imports the configerator changes from D35613093 and adds support on Mononoke to changing the deleted manifest version.

It changes a bit from the approach I was using previously, where we had `deleted_manifest` name for V1 and `deleted_manifest2` for V2. Now it uses the same approach blame used, which is having the same name for both, but always changing the underlying implementation based on the config.

Since there was not much usage of deleted manifests directly (only on 3 files), I went with checking the config and using static types on each of them, instead of adding a wrapper type around DMv1 or v2, as that would involve creating a lot of duplicated code.

Reviewed By: kris1319

Differential Revision: D35814371

fbshipit-source-id: 4edf87296e2ea9986f9af18cb2a7e6914ea5ffbe