testlib: load dynamicconfig
Summary:
We want the tests to run with production configs as much as possible.
Let's allow them to load the non-remote portions of dynamicconfig. This should
also help encourage people to move configs from remote-configuration to the
static dynamicconfig.

There is a bit of an issue right now where the static dynamicconfig is fb-only.
So these tests may behave differently when run against an open source build. We
probably need to move a bunch of the fb-only configs to be public configs.

Reviewed By: quark-zju

Differential Revision: D35820978

fbshipit-source-id: 24c8fe73198244b99e0beeae22b1e93cdccb9872