add --press-to-continue flag to edenfsctl commands
Summary:
When an `edenfsctl` command is invoked from the E-Menu, a PowerShell Window is opened and is immediately closed once the command finishes. This makes it hard/impossible for the user to see the command's output.

The desired behavior here is to keep the Window open long enough for the user to acknowledge the output. I did this by adding a `--press-to-continue` flag to all `edenfsctl` commands. This will mainly be used by the E-Menu, but may have uses elsewhere.

Reviewed By: chadaustin

Differential Revision: D36137709

fbshipit-source-id: a46bebb5bb5820949cd4b70579a29406afb34048