Do not allow non-found git-lfs files in normal mega_grepo_sync opreations.
Summary:
Do not allow non-found git-lfs files in normal mega_grepo_sync opreations.
Note that this can be overriden via the command line and it is still the default when running the stand alone branch_forest_updated program.

The effect of this is that we will now fail to sync branch-forest projects if the git-lfs content gets a "not found" response from the server, and throw an error. The follow on effect is that our sync-service will try to sync it next time the branch comes around again and repeat.
This is different from the current behavior where it will just accept that it cannot be found and put in the git-lfs text pointer instead. But we have seen this happening quite often lately, likely due to some latency from uploading to git-lfs to when it is avalible in the read-URL/server.

Differential Revision: D36150297

fbshipit-source-id: c049676049850db9e97fe84e99f7d47c5e35eea8