windows: fix StartupLogger on Windows
Summary:
This diff enables StartupLogger on Windows.

In the past we disabled the StartupLogger on Windows because it will cause the daemon to terminate when the user closes their current terminal, or it would cause us to not being able to capture output from subprocess we spawn.

This diff works around the issue by redirecting the output from child process's stderr to a pipe. Then creating a thread in the parent process to read and print it.

Once the startup process finishes, the parent process will terminate, and the stderr handle used by the daemon process would be closed.

Reviewed By: xavierd

Differential Revision: D35833750

fbshipit-source-id: 71d01b20da8219cf6b186e7c46590fbfb9d14623