model: remove an unecessary std::move in Tree:tryDeserialize
Summary:
The caller of this function wants to return a std::unique_ptr, thus instead of
using an optional to denote failure to deserialize, we can use a nullptr to
encode the same information. This prevents the caller from moving the Tree into
a unique_ptr if the deserialization succeded.

Reviewed By: genevievehelsel

Differential Revision: D36179488

fbshipit-source-id: 735fd6480faa8a520e538d065c68996cc3bb7615