diff: allow reporting directory changes too
Summary:
The core diff code was only able to report changes as desired by Mercurial, but
this code is evolving to support reporting changes to Watchman, which needs to
be able to report directory changes. As a first step, let's have the
DiffCallback methods take a dtype_t as an argument and call these with
directories.

Reviewed By: chadaustin

Differential Revision: D35985932

fbshipit-source-id: a9c20cd086c358903b21321162ee9c9fa657dfc7