Temporarily ignore custom key prefix on DMv2
Summary:
Custom key prefixes are done by derived data name, which is the same for DMv1 and DMv2. Since we're deriving DMv2 for the first time, we don't need any prefix, so let's ignore it.

Without this diff, we would backfill/use dmv2 with "deletedmanifestroot2.r2.XXX" unnecessarily, which is not how I've been backfilling so far, so this diff is needed to be able to change the clients.

Once we move to DMv2, we can remove the prefix from the config, and then change this code to consider the prefix again, so that we can add it in the future if necessary.

Reviewed By: mitrandir77

Differential Revision: D36170347

fbshipit-source-id: 65556d35e084fa956e227dfda09c98b6d693fb92