pull: workaround multiple heads lazy pull in fbsource
Summary:
The existing assumes `heads(mastergroup())` to be just one single commit
matching the main bookmark (`remote/master`). That's no longer the case with
some failed merges as heads. Let's make the code aware of the case and try to
pick a better "oldhead" for the lazy pull to work.

As we're here, update the hardcoded "master:" to the actual bookmark name.

Reviewed By: DurhamG

Differential Revision: D36214778

fbshipit-source-id: 5209e114eff7fe788249c0b8489938dee0329656