call x2pagentctl rage instead of doing it ourselves
Summary:
Previously we did it ourselves - we had to keep paths to logs updated, and we had to know how to get status of the agent. Not it's handled by the team who owns x2pagent!

I could also remove `-n` but that won't work off-VPN because it needs to access intern to make a paste. (`n` stores the rage in a paste).

Reviewed By: HarveyHunt

Differential Revision: D36160669

fbshipit-source-id: 01a6037e55ee7246a866087bb1061d18bf950487