rage: remove color escapes from hg status section
Summary:
Non-windows users were showing color escape characters in their rage reports, which made them somewhat hard to read. This fixes that.

The cause for this diff was that in the recent rewrite of the `hg status` command we would only check if it was possible to output colors in the current terminal, instead of checking for the `ui.color` like the the rest of our Python commands.

Differential Revision: D36079055

fbshipit-source-id: 1fdca8a1292bdff9f16e3dc58d333b4407a60212