add manifest to cmake build
Summary:
Cmake automatically includes its own manifest in executables. We should make CMake aware of our custom manifest so that it combines them together to avoid a "duplicate manifest" error

In addition, we should only include the manifest in the resource file if we're building with buck. The Cmake build will break if we try to include the manifest in the resource file. This is because every target that includes notifications will pull in the resource file, and this causes all of those targets to have duplicated manifest errors during linking.

Reviewed By: chadaustin

Differential Revision: D36318044

fbshipit-source-id: dde35e5602f0161c4961103970b880f98e399098