Fix a crash in SqliteStatement
Summary:
We were seeing a crash on Windows and this diff fixes that.

`sqlite3_reset` will return the same error code as it did in `sqlite3_step` if there is any error happened during that statement execution.

This was fine previously because the exception was handled when we call `sqlite3_step`. However after my previous diff, we are calling `sqlite3_reset` in the newly introduced scope guard.

This means we will be rethrowing the same exception in the destructor, causing EdenFS to crash.

Reviewed By: xavierd

Differential Revision: D36264955

fbshipit-source-id: 26322307b366ba637431144ee132ca82f894f21f