model: move Tree closer to the PathMap API
Summary:
With Tree bound to be backed by PathMap in the near future (see previous diff
for why), we need to unify their API to ease this transition.

One drawback of this diff is that we now store the PathComponent twice: once in
the std::vector, and once in the TreeEntry. In a future diff, the PathComponent
from TreeEntry will be removed, solving this issue.

Reviewed By: chadaustin

Differential Revision: D36215509

fbshipit-source-id: 12390c825ef706f13481d832f255a5a0d8770717