repolock: add Rust repo.wlock() equivalent
Summary:
Add repolock::lock_working_copy() which acquires the "wlock" lock and writes {hostname}:{pid} as the contents.

Python has more complicated logic for the lock contents - I did the absolute minimum for compatibility. In the future we should switch the lock contents to be the runlog id for the command.

Reviewed By: quark-zju

Differential Revision: D36016276

fbshipit-source-id: 7eff93111f8607748eb6486f7c0ad9dc5b8bb844