config: move ui.timeout defaults to rust
Summary:
I'm going to use these in rust repo lock acquisition loop (a la Python).

Note that I removed the ui.timeout.warn default value. This default of 0 conflicted with the "None" default value the code used when fetching this config value. The effective default was "None", so simplify things by just not setting a default value in the builtin config (and not specifying a default in the code when fetching the value).

Reviewed By: quark-zju

Differential Revision: D36016273

fbshipit-source-id: 418c316409fea5a0c8fd55d36d9f36d6e03605af